tijiao

+++ src/main/java/com/zyh/controller/classes/ClassController.java
@@ -1,5 +1,6 @@
 package com.zyh.controller.classes;
 
+import java.util.Date;
 import java.util.List;
 
 import org.apache.log4j.Logger;
@@ -8,6 +9,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.zyh.entity.classcourse.ZyhClassCourse;
 import com.zyh.entity.classcourse.ZyhClassCourseExample;
@@ -18,6 +20,7 @@ import com.zyh.service.classcourse.IClassCourseService;
 import com.zyh.service.classteacher.IClassTeacherService;
 
 @RestController
+@RequestMapping("/teachcourse")
 public class ClassController {
 	private Logger log = Logger.getLogger("error");
 	
@@ -39,6 +42,7 @@ public class ClassController {
 		try {
 			ZyhClassTeacher teacher = om.readValue(json, ZyhClassTeacher.class);
 			if (null != teacher) {
+				teacher.setCreatetime(new Date());
 				classTeacherService.addClassTeacher(teacher);
 				responeToWeb.setMsg("新增成功");
 				responeToWeb.setSuccess(true);
@@ -66,7 +70,7 @@ public class ClassController {
 		ObjectMapper om = new ObjectMapper();
 		try {
 			ZyhClassTeacher teacher = om.readValue(json, ZyhClassTeacher.class);
-			if (null != teacher) {
+			if (null != teacher && null!=teacher.getId() && !"".equals(teacher.getId())) {
 				classTeacherService.updateClassTeacher(teacher);
 				responeToWeb.setMsg("修改成功");
 				responeToWeb.setSuccess(true);
@@ -89,7 +93,8 @@ public class ClassController {
 		ResponeToWeb responeToWeb = new ResponeToWeb();
 		ObjectMapper om = new ObjectMapper();
 		try {
-			String id = om.readValue(json, String.class);
+			JsonNode node = om.readTree(json);
+			String id =  node.get("id").asText();
 			if (null != id && !"".equals(id)) {
 				ZyhClassCourseExample example = new ZyhClassCourseExample();
 				Criteria criteria =example.createCriteria();
@@ -128,7 +133,10 @@ public class ClassController {
 		ObjectMapper om = new ObjectMapper();
 		try {
 			ZyhClassCourse course = om.readValue(json, ZyhClassCourse.class);
-			if (null != course) {
+			//教师必填
+			if (null != course && null!=course.getTeacherid() && 
+					!"".equals(course.getTeacherid())) {
+				course.setCreatetime(new Date());
 				classCourseService.addClassCourse(course);
 				responeToWeb.setMsg("新增成功");
 				responeToWeb.setSuccess(true);
@@ -156,7 +164,7 @@ public class ClassController {
 		ObjectMapper om = new ObjectMapper();
 		try {
 			ZyhClassCourse course = om.readValue(json, ZyhClassCourse.class);
-			if (null != course) {
+			if (null != course && null!=course.getId() && !"".equals(course.getId())) {
 				classCourseService.updateClassCourse(course);
 				responeToWeb.setMsg("修改成功");
 				responeToWeb.setSuccess(true);
@@ -179,8 +187,9 @@ public class ClassController {
 		ResponeToWeb responeToWeb = new ResponeToWeb();
 		ObjectMapper om = new ObjectMapper();
 		try {
-			String id = om.readValue(json, String.class);
-			if (null != id) {
+			JsonNode node = om.readTree(json);
+			String id =  node.get("id").asText();
+			if (null != id && !"".equals(id)) {
 				classCourseService.deleteClassCourse(id);
 				responeToWeb.setMsg("删除成功");
 				responeToWeb.setSuccess(true);
