+++ src/main/java/com/zyh/controller/policy/PolicyController.java
@@ -1,5 +1,6 @@
 package com.zyh.controller.policy;
 
+import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -10,15 +11,16 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.zyh.entity.common.ResponeToWeb;
-import com.zyh.entity.news.ZyhNews;
 import com.zyh.entity.policy.ZyhPolicy;
 import com.zyh.entity.policy.ZyhPolicyExample;
 import com.zyh.entity.policy.ZyhPolicyExample.Criteria;
 import com.zyh.service.policy.IPolicyService;
 
 @RestController
+@RequestMapping("/policy")
 public class PolicyController {
 
 	private Logger log = Logger.getLogger("error");
@@ -32,13 +34,19 @@ public class PolicyController {
 		ObjectMapper om = new ObjectMapper();
 		Map<String, Object> map = new HashMap<String, Object>();
 		try {
-			String policyid = om.readValue(json, String.class);
-			if (null != policyid ) {
-				ZyhPolicy policy = policyService.findPolicyById(policyid);
-				map.put("result", policy);
-				responeToWeb.setMsg("查询成功");
-				responeToWeb.setSuccess(true);
-				responeToWeb.setValue(map);
+			JsonNode node = om.readTree(json);
+			String policyid =  node.get("policyid").asText();
+			if (null != policyid && !"".equals(policyid)) {
+				ZyhPolicy policy = policyService.queryPolicyById(policyid);
+				if(null == policy ){
+					responeToWeb.setMsg("查询失败,信息缺失");
+					responeToWeb.setSuccess(false);
+				}else{
+					map.put("result", policy);
+					responeToWeb.setMsg("查询成功");
+					responeToWeb.setSuccess(true);
+					responeToWeb.setValue(map);
+				}
 			} else {
 				responeToWeb.setMsg("查询失败,信息缺失");
 				responeToWeb.setSuccess(false);
@@ -59,12 +67,19 @@ public class PolicyController {
 		Map<String, Object> map = new HashMap<String, Object>();
 		try {
 			//标题查询
-			String title = om.readValue(json, String.class);
+			String title ="";
+			if(null!=json && !"".equals(json)){
+				JsonNode node = om.readTree(json);
+				if(null!=node.get("title") && !"".equals(node.get("title"))){
+					title =  node.get("title").asText();
+				}
+			}
 			ZyhPolicyExample zyhPolicyExample = new ZyhPolicyExample();
 			zyhPolicyExample.setOrderByClause("pubtime desc");
+			Criteria criteria = zyhPolicyExample.createCriteria();
+			criteria.andIfgroundEqualTo("1");
 			if(null!=title || "" !=title){
-				Criteria criteria = zyhPolicyExample.createCriteria();
-				criteria.andTitleLike(title);
+				criteria.andTitleLike("%"+title+"%");
 			}
 			List<ZyhPolicy> policylist = policyService.findPolicyList(zyhPolicyExample);
 			map.put("result", policylist);
@@ -91,6 +106,7 @@ public class PolicyController {
 		try {
 			ZyhPolicy policy = om.readValue(json, ZyhPolicy.class);
 			if (null != policy) {
+				policy.setPubtime(new Date());
 				policyService.addPolicy(policy);
 				responeToWeb.setMsg("新增成功");
 				responeToWeb.setSuccess(true);
@@ -118,7 +134,7 @@ public class PolicyController {
 		ObjectMapper om = new ObjectMapper();
 		try {
 			ZyhPolicy policy = om.readValue(json, ZyhPolicy.class);
-			if (null != policy) {
+			if (null != policy && null!=policy.getId() && !"".equals(policy.getId())) {
 				policyService.updatePolicy(policy);
 				responeToWeb.setMsg("修改成功");
 				responeToWeb.setSuccess(true);
@@ -141,8 +157,9 @@ public class PolicyController {
 		ResponeToWeb responeToWeb = new ResponeToWeb();
 		ObjectMapper om = new ObjectMapper();
 		try {
-			String id = om.readValue(json, String.class);
-			if (null != id) {
+			JsonNode node = om.readTree(json);
+			String id =  node.get("id").asText();
+			if (null != id && !"".equals(id)) {
 				policyService.deletePolicy(id);
 				responeToWeb.setMsg("删除成功");
 				responeToWeb.setSuccess(true);
